rename class
