+++ android/app/src/main/java/nifty/intern/teamc/database/DatabaseManager.java
@@ -1,7 +1,6 @@
-package nifty.intern.teamc.beaconreminder;
+package nifty.intern.teamc.database;
 
 import android.content.Context;
-import android.util.StringBuilderPrinter;
 
 import com.nifty.cloud.mb.core.DoneCallback;
 import com.nifty.cloud.mb.core.NCMB;
@@ -11,8 +10,8 @@ import com.nifty.cloud.mb.core.NCMBObject;
 /**
  * Created by USER on 2016/08/31.
  */
-public class Mbaas {
-    private static final String APP_KEY="APP_KEY";
+public class DatabaseManager {
+    private static final String APP_KEY="APP_KE";
     private static final String CLIENT_KEY="CLIENT_KEY";
     public static void initialize(Context context) {
         NCMB.initialize(context, APP_KEY, CLIENT_KEY);
@@ -31,13 +30,12 @@ public class Mbaas {
             public void done(NCMBException e) {
                 if(e != null){
                     //保存に失敗した場合の処理
-                    System.out.println("failed");
+                    System.out.println("storeTask failed");
                 }else {
                     //保存に成功した場合の処理
-                    System.out.println("success");
+                    System.out.println("storeTask success");
                 }
             }
         });
     }
-
 }
